refactor: moving component to client
