+++ apps/web/messages/vi.json
@@ -13,7 +13,17 @@
     "contact-sales-desc": "Nếu bạn có bất kỳ câu hỏi nào về gói đăng ký của mình hoặc muốn nâng cấp lên gói doanh nghiệp, vui lòng liên hệ với bộ phận bán hàng của chúng tôi.",
      "plan-desc":"* Gói đăng ký của bạn sẽ có hiệu lực ngay lập tức và sẽ tự động gia hạn hàng tháng hoặc hàng năm, tùy thuộc vào gói bạn chọn.",
     "subscription-history": "Lịch sử đăng ký",
-    "subscription-history-desc": "Không có lịch sử đăng ký nào."
+    "subscription-history-desc": "Không có lịch sử đăng ký nào.",
+    "success":{
+      "title": "Thanh toán thành công",
+      "message": "Cảm ơn bạn đã thanh toán. Gói đăng ký của bạn đã được cập nhật thành công.",
+      "payment-summary" : "Chi tiết thanh toán",
+       "plan": "Gói",
+       "paid": "Đã thanh toán",
+       "status": "Trạng thái",
+       "payment-method": "Phương thức thanh toán",
+       "invoice-id": "ID hoá đơn"
+    }
   },
   "common": {
     "success": "Thành công",
