+++ packages/types/src/supabase.ts
@@ -7783,48 +7783,48 @@ export type Database = {
       };
       count_search_users: {
         Args:
+          | { search_query: string }
           | {
-              enabled_filter?: boolean;
               search_query: string;
               role_filter?: string;
-            }
-          | { search_query: string };
+              enabled_filter?: boolean;
+            };
         Returns: number;
       };
       create_ai_chat: {
-        Args: { message: string; title: string; model: string };
+        Args: { title: string; message: string; model: string };
         Returns: string;
       };
       generate_cross_app_token: {
         Args:
           | {
               p_user_id: string;
-              p_expiry_seconds?: number;
               p_origin_app: string;
               p_target_app: string;
+              p_expiry_seconds?: number;
             }
           | {
               p_user_id: string;
-              p_session_data?: Json;
-              p_expiry_seconds?: number;
-              p_target_app: string;
               p_origin_app: string;
+              p_target_app: string;
+              p_expiry_seconds?: number;
+              p_session_data?: Json;
             };
         Returns: string;
       };
       get_challenge_stats: {
-        Args: { user_id_param: string; challenge_id_param: string };
+        Args: { challenge_id_param: string; user_id_param: string };
         Returns: {
           total_score: number;
           problems_attempted: number;
         }[];
       };
       get_daily_income_expense: {
-        Args: { past_days?: number; _ws_id: string };
+        Args: { _ws_id: string; past_days?: number };
         Returns: {
-          total_expense: number;
-          total_income: number;
           day: string;
+          total_income: number;
+          total_expense: number;
         }[];
       };
       get_daily_prompt_completion_tokens: {
@@ -7879,9 +7879,8 @@ export type Database = {
           _has_unit?: boolean;
         };
         Returns: {
-          created_at: string;
-          name: string;
           id: string;
+          name: string;
           manufacturer: string;
           unit: string;
           unit_id: string;
@@ -7889,6 +7888,7 @@ export type Database = {
           price: number;
           amount: number;
           ws_id: string;
+          created_at: string;
         }[];
       };
       get_inventory_products_count: {
@@ -7928,7 +7928,7 @@ export type Database = {
         Returns: number;
       };
       get_possible_excluded_groups: {
-        Args: { included_groups: string[]; _ws_id: string };
+        Args: { _ws_id: string; included_groups: string[] };
         Returns: {
           id: string;
           name: string;
@@ -7948,11 +7948,11 @@ export type Database = {
       get_session_statistics: {
         Args: Record<PropertyKey, never>;
         Returns: {
-          latest_session_date: string;
+          total_count: number;
           unique_users_count: number;
           active_count: number;
           completed_count: number;
-          total_count: number;
+          latest_session_date: string;
         }[];
       };
       get_session_templates: {
@@ -7962,9 +7962,9 @@ export type Database = {
           limit_count?: number;
         };
         Returns: {
-          category_id: string;
           title: string;
           description: string;
+          category_id: string;
           task_id: string;
           tags: string[];
           category_name: string;
@@ -7978,20 +7978,20 @@ export type Database = {
       get_submission_statistics: {
         Args: Record<PropertyKey, never>;
         Returns: {
-          unique_users_count: number;
-          latest_submission_date: string;
           total_count: number;
+          latest_submission_date: string;
+          unique_users_count: number;
         }[];
       };
       get_transaction_categories_with_amount: {
         Args: Record<PropertyKey, never>;
         Returns: {
-          ws_id: string;
-          amount: number;
           id: string;
           name: string;
           is_expense: boolean;
+          ws_id: string;
           created_at: string;
+          amount: number;
         }[];
       };
       get_user_role: {
@@ -8001,44 +8001,44 @@ export type Database = {
       get_user_session_stats: {
         Args: { user_id: string };
         Returns: {
-          active_sessions: number;
           total_sessions: number;
+          active_sessions: number;
           current_session_age: unknown;
         }[];
       };
       get_user_sessions: {
         Args: { user_id: string };
         Returns: {
-          ip: string;
-          updated_at: string;
-          user_agent: string;
-          is_current: boolean;
           session_id: string;
           created_at: string;
+          updated_at: string;
+          user_agent: string;
+          ip: string;
+          is_current: boolean;
         }[];
       };
       get_user_tasks: {
         Args: { _board_id: string };
         Returns: {
-          list_id: string;
-          completed: boolean;
-          board_id: string;
-          name: string;
-          start_date: string;
-          description: string;
-          end_date: string;
-          priority: number;
           id: string;
+          name: string;
+          description: string;
+          priority: number;
+          completed: boolean;
+          start_date: string;
+          end_date: string;
+          list_id: string;
+          board_id: string;
         }[];
       };
       get_user_whitelist_status: {
         Args: { user_id_param: string };
         Returns: {
           is_whitelisted: boolean;
-          allow_role_management: boolean;
+          enabled: boolean;
           allow_challenge_management: boolean;
           allow_manage_all_challenges: boolean;
-          enabled: boolean;
+          allow_role_management: boolean;
         }[];
       };
       get_workspace_drive_size: {
@@ -8054,24 +8054,24 @@ export type Database = {
         Returns: number;
       };
       get_workspace_transactions_count: {
-        Args: { end_date?: string; start_date?: string; ws_id: string };
+        Args: { ws_id: string; start_date?: string; end_date?: string };
         Returns: number;
       };
       get_workspace_user_groups: {
         Args: {
-          search_query: string;
-          excluded_tags: string[];
-          included_tags: string[];
           _ws_id: string;
+          included_tags: string[];
+          excluded_tags: string[];
+          search_query: string;
         };
         Returns: {
+          id: string;
+          name: string;
+          notes: string;
+          ws_id: string;
+          tags: string[];
           tag_count: number;
           created_at: string;
-          tags: string[];
-          ws_id: string;
-          notes: string;
-          name: string;
-          id: string;
         }[];
       };
       get_workspace_user_groups_count: {
@@ -8086,7 +8086,6 @@ export type Database = {
           search_query: string;
         };
         Returns: {
-          group_count: number;
           id: string;
           avatar_url: string;
           full_name: string;
@@ -8103,6 +8102,7 @@ export type Database = {
           balance: number;
           ws_id: string;
           groups: string[];
+          group_count: number;
           linked_users: Json;
           created_at: string;
           updated_at: string;
@@ -8117,15 +8117,15 @@ export type Database = {
         Returns: number;
       };
       get_workspace_wallets_expense: {
-        Args: { start_date?: string; ws_id: string; end_date?: string };
+        Args: { ws_id: string; start_date?: string; end_date?: string };
         Returns: number;
       };
       get_workspace_wallets_income: {
-        Args: { start_date?: string; end_date?: string; ws_id: string };
+        Args: { ws_id: string; start_date?: string; end_date?: string };
         Returns: number;
       };
       has_other_owner: {
-        Args: { _user_id: string; _ws_id: string };
+        Args: { _ws_id: string; _user_id: string };
         Returns: boolean;
       };
       insert_ai_chat_message: {
@@ -8137,7 +8137,7 @@ export type Database = {
         Returns: boolean;
       };
       is_member_invited: {
-        Args: { _org_id: string; _user_id: string };
+        Args: { _user_id: string; _org_id: string };
         Returns: boolean;
       };
       is_nova_challenge_manager: {
@@ -8153,7 +8153,7 @@ export type Database = {
         Returns: boolean;
       };
       is_nova_user_id_in_team: {
-        Args: { _team_id: string; _user_id: string };
+        Args: { _user_id: string; _team_id: string };
         Returns: boolean;
       };
       is_org_member: {
@@ -8189,11 +8189,11 @@ export type Database = {
         Returns: Json;
       };
       nova_get_user_daily_sessions: {
-        Args: { user_id: string; challenge_id: string };
+        Args: { challenge_id: string; user_id: string };
         Returns: number;
       };
       nova_get_user_total_sessions: {
-        Args: { user_id: string; challenge_id: string };
+        Args: { challenge_id: string; user_id: string };
         Returns: number;
       };
       remove_board_tags: {
@@ -8209,19 +8209,19 @@ export type Database = {
         Returns: number;
       };
       revoke_user_session: {
-        Args: { session_id: string; target_user_id: string };
+        Args: { target_user_id: string; session_id: string };
         Returns: boolean;
       };
       search_boards_by_tags: {
         Args: {
-          search_tags: string[];
           workspace_id: string;
+          search_tags: string[];
           match_all?: boolean;
         };
         Returns: {
+          board_id: string;
           board_name: string;
           board_tags: Json;
-          board_id: string;
         }[];
       };
       search_users: {
@@ -8235,6 +8235,13 @@ export type Database = {
               enabled_filter?: boolean;
             };
         Returns: {
+          id: string;
+          display_name: string;
+          deleted: boolean;
+          avatar_url: string;
+          handle: string;
+          bio: string;
+          created_at: string;
           user_id: string;
           enabled: boolean;
           allow_challenge_management: boolean;
@@ -8244,27 +8251,20 @@ export type Database = {
           new_email: string;
           birthday: string;
           team_name: string[];
-          id: string;
-          display_name: string;
-          deleted: boolean;
-          avatar_url: string;
-          handle: string;
-          bio: string;
-          created_at: string;
         }[];
       };
       search_users_by_name: {
         Args: {
-          result_limit?: number;
           search_query: string;
+          result_limit?: number;
           min_similarity?: number;
         };
         Returns: {
-          relevance: number;
+          id: string;
           handle: string;
           display_name: string;
           avatar_url: string;
-          id: string;
+          relevance: number;
         }[];
       };
       sum_quiz_scores: {
@@ -8286,7 +8286,7 @@ export type Database = {
         Returns: undefined;
       };
       update_session_total_score: {
-        Args: { user_id_param: string; challenge_id_param: string };
+        Args: { challenge_id_param: string; user_id_param: string };
         Returns: undefined;
       };
       validate_and_normalize_board_tags: {
@@ -8298,14 +8298,14 @@ export type Database = {
         Returns: boolean;
       };
       validate_cross_app_token: {
-        Args: { p_target_app: string; p_token: string };
+        Args: { p_token: string; p_target_app: string };
         Returns: string;
       };
       validate_cross_app_token_with_session: {
-        Args: { p_target_app: string; p_token: string };
+        Args: { p_token: string; p_target_app: string };
         Returns: {
-          session_data: Json;
           user_id: string;
+          session_data: Json;
         }[];
       };
     };
